extracted progress content from project content
